Rename files enginefinder* to operationgraph_descriptor*
