+++ src/CMakeLists.txt
@@ -122,8 +122,8 @@ set( MIOpen_Source
     fusion/problem_description.cpp
     generic_search.cpp
     graphapi/convolution.cpp
-    graphapi/enginefinder.cpp
     graphapi/graphapi.cpp
+    graphapi/operationgraph_descriptor.cpp
     graphapi/opgraph.cpp
     graphapi/matmul.cpp
     graphapi/pointwise.cpp
