+++ src/graphapi/graphapi.cpp
@@ -25,8 +25,8 @@
  *******************************************************************************/
 #include <miopen/errors.hpp>
 #include <miopen/graphapi/convolution.hpp>
-#include <miopen/graphapi/enginefinder.hpp>
 #include <miopen/graphapi/graphapi.hpp>
+#include <miopen/graphapi/operationgraph_descriptor.hpp>
 #include <miopen/graphapi/pointwise.hpp>
 #include <miopen/graphapi/reduction.hpp>
 #include <miopen/graphapi/rng.hpp>
